move shared election logic
