+++ src/geoscore_de/data_flow/features/election_21.py
@@ -3,9 +3,8 @@ from pathlib import Path
 
 import pandas as pd
 
-from geoscore_de.data_flow.features.base import BaseFeature
+from geoscore_de.data_flow.features.election import BaseElectionFeature
 from geoscore_de.data_flow.features.municipality import DEFAULT_RAW_DATA_PATH as MUNICIPALITY_RAW_DATA_PATH
-from geoscore_de.data_flow.features.municipality import MunicipalityFeature
 from geoscore_de.data_flow.features.utils import load_election_zip, move_extracted_file
 
 ZIP_URL = "https://www.bundeswahlleiterin.de/en/dam/jcr/c2cd99e6-064e-4ebc-b634-f86b5c0e14b3/btw21_wbz.zip"
@@ -16,7 +15,7 @@ HAMBURG_CITY_AGS = "02000000"
 BERLIN_CITY_AGS = "11000000"
 
 
-class Election21Feature(BaseFeature):
+class Election21Feature(BaseElectionFeature):
     """Feature class for election 2021 data."""
 
     def __init__(
@@ -28,7 +27,7 @@ class Election21Feature(BaseFeature):
         fix_missing: bool = True,
         **kwargs,
     ):
-        super().__init__(**kwargs)
+        super().__init__(municipality_data_path=municipality_data_path, **kwargs)
         self.url = url
         self.raw_data_path = raw_data_path
         self.tform_data_path = tform_data_path
@@ -103,42 +102,6 @@ class Election21Feature(BaseFeature):
         normalized_ags = normalized_ags.where(~berlin_district_mask, BERLIN_CITY_AGS)
         return normalized_ags
 
-    def _fix_missing(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Municipalities that are missing in the election data but present in the municipality data
-        are filled using einschl. keyword in the municipality name. If a municipality is missing
-        but is included in another municipality's data add the missing municipality with the same values
-        as the including municipality.
-
-        Args:
-            df (pd.DataFrame): DataFrame containing the election data.
-
-        Returns:
-            pd.DataFrame: DataFrame with added rows for missing municipalities.
-        """
-        municipality_feature = MunicipalityFeature(self.municipality_data_path)
-        df_muni = municipality_feature.load()[["AGS", "Municipality"]]
-
-        df_merged = df.merge(df_muni, on="AGS", how="outer", indicator=True)
-
-        missing_munis = df_merged[df_merged["_merge"] == "right_only"]
-        einschl_munis = df_merged[df_merged["Gemeinde Name"].str.contains("einschl.", na=False)]
-
-        # iterate over missing municipalities and check if their name is included in any of the einschl. municipalities
-        for _, missing_row in missing_munis.iterrows():
-            missing_name = missing_row["Municipality"]
-            # Find first matching einschl. municipality
-            for _, einschl_row in einschl_munis.iterrows():
-                einschl_name = einschl_row["Gemeinde Name"]
-                if missing_name in einschl_name:
-                    # If the missing municipality is included in the einschl. municipality
-                    # add row to original df with the same values as the einschl. municipality
-                    # except for AGS
-                    new_row = df[df["AGS"] == einschl_row["AGS"]].copy()
-                    new_row["AGS"] = missing_row["AGS"]
-                    df = pd.concat([df, new_row], ignore_index=True)
-                    break
-        return df
-
     def transform(self, df: pd.DataFrame) -> pd.DataFrame:
         """Transform raw election 21 data to include only relevant columns.
 
@@ -169,7 +132,7 @@ class Election21Feature(BaseFeature):
 
         # fix einschl.
         if self.fix_missing:
-            df = self._fix_missing(df)
+            df = self._fix_missing(df, muni_name_col="Gemeinde Name")
 
         df = df.drop(columns=["Gemeinde Name"], errors="ignore")
